feat: disable sidekiq startup "enqueues" on 90% of processes
ref: https://app.asana.com/0/1198103229139841/1200127065515039

Restarting sidekiq in production currently leads to a huge CPU
spike on redis as hundreds of sidekiq processes simultaneously
spawn and look for jobs to schedule. This is technically all
redundant after the first process, however, it is risky to rely
only on one process (what if the process crashes or the server
goes down).

The suggested temporary patch (implemented in this commit)
randomizes which processes will start enqueuing jobs on startup,
allowing only 10% of calls to proceed. This should reduce the load
load on redis for now for very minimal work, but long-term we
should look at not having to rely on custom forks of sidekiq and
sidekiq-cron to do our scheduling.